refactor(phase-2m): extract OpenProjectRecordsService (CRUD)
Continues the openproject_client.py god-class decomposition (ADR-002).
The six generic ActiveRecord CRUD helpers move into a focused
``OpenProjectRecordsService`` exposed as ``self.records``.
``OpenProjectClient`` keeps a thin delegator for each method so
existing callers work unchanged.

Methods moved
-------------
- ``find_record`` (id or conditions hash)
- ``batch_find_records`` — ``@batch_idempotent`` decorator travels
  with the moved method, and the signature gains a keyword-only
  ``headers`` kwarg (consistent with the user/project services)
- ``create_record``
- ``update_record``
- ``delete_record``
- ``find_all_records`` (where + includes + limit, file-based
  read for multi-record results)

What stays on the client
------------------------
- ``_retry_with_exponential_backoff`` — used by 4 sibling services
  (records, custom_fields, users, projects); keeping it on the
  client avoids cross-service references.
- ``_validate_batch_size``, ``_validate_model_name`` (instance method
  + module-level function), ``_build_safe_batch_query``,
  ``escape_ruby_single_quoted`` — generic helpers shared across
  services. The records service reaches them via
  ``self._client.<helper>`` or lazy imports.

Numbers
-------
- ``openproject_client.py``: 4,685 → 4,350 LOC (-335)
- ``openproject_records_service.py``: 0 → 525 LOC (new)
- Cumulative across phases 2a-2m: ``openproject_client.py``
  7,342 → 4,350 LOC (-2,992, -40.8%)

Verification
------------
- pytest tests/unit: 953 passed
- mypy src/: clean (105 files)
- ruff check / ruff format: clean
